big upgrate

+++ Coords.cpp
@@ -5,17 +5,17 @@
 #include "Coords.h"
 #include <iostream>
 
-Coords::Coords(int x, int y) : x(x), y(y) {}
+Coords::Coords(size_t x, size_t y) : x(x), y(y) {}
 
-std::pair<int, int> Coords::getXY() const {
+std::pair<size_t , size_t> Coords::getXY() const {
     return std::make_pair(x,y);
 }
 
-int Coords::getX() const {
+size_t Coords::getX() const {
     return x;
 }
 
-int Coords::getY() const {
+size_t Coords::getY() const {
     return y;
 }
